Fixed some more warnings and resolved deprecated methods

+++ org.exist.eclipse.xquery.core/src/org/w3c/xqparser/XPathTokenManager.java
@@ -343,7 +343,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -422,7 +422,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -522,7 +522,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -612,7 +612,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -739,7 +739,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -873,7 +873,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -1018,7 +1018,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -1194,7 +1194,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -1267,7 +1267,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -1332,7 +1332,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -1458,7 +1458,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -1527,7 +1527,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -1592,7 +1592,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -1964,7 +1964,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -2033,7 +2033,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -2098,7 +2098,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -2163,7 +2163,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -2244,7 +2244,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -2361,7 +2361,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -2426,7 +2426,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -2491,7 +2491,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -2647,7 +2647,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -2737,7 +2737,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -2802,7 +2802,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -2867,7 +2867,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -2932,7 +2932,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -5101,7 +5101,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -5412,7 +5412,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -5481,7 +5481,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -5715,7 +5715,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -5784,7 +5784,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -5849,7 +5849,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
@@ -5914,7 +5914,7 @@ public class XPathTokenManager implements XPathConstants {
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
